trQr Social Media Websites
trQr (pronounce TRACKER) is a social media website concept that offers
users to track their progress. The goal is to allow users to make posts
with text and photos to show their development throughout the trQ. Users
will be able to share their trQs in different categories: public, all
friends, some friends, or no one. When setting up a public trQ, there
will be an option to add a hashtag (#). When hash-tagging a trQ, others
will be able to use the same trQ format that the original user set up.
Finally, friends will be able to make comments about other people's
trQs.